rename: package 이동
- toss.exception에 예외 관련 클래스 이동
