+++ src/main/java/roomescape/payment/service/PaymentService.java
@@ -3,7 +3,7 @@ package roomescape.payment.service;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import roomescape.payment.TossPaymentClient;
+import roomescape.payment.toss.TossPaymentClient;
 import roomescape.payment.domain.Payment;
 import roomescape.payment.dto.TossPaymentRequest;
 import roomescape.payment.dto.TossPaymentResponse;
