+++ src/test/java/roomescape/payment/FakePaymentRestClientConfig.java
@@ -3,6 +3,7 @@ package roomescape.payment;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Profile;
+import roomescape.payment.toss.TossPaymentClient;
 
 @TestConfiguration
 @Profile("test")
