+++ src/test/java/roomescape/payment/FakeTossPaymentClient.java
@@ -3,6 +3,7 @@ package roomescape.payment;
 import roomescape.payment.domain.PaymentType;
 import roomescape.payment.dto.TossPaymentRequest;
 import roomescape.payment.dto.TossPaymentResponse;
+import roomescape.payment.toss.TossPaymentClient;
 
 public class FakeTossPaymentClient implements TossPaymentClient {
